+++ test/integration/routing/sla_caches_test.rb
@@ -42,6 +42,8 @@ class RoutingSlaCachesTest < ApplicationSlaRoutingTestCase
     should_route 'GET /sla/caches/1/refresh' => 'sla_caches#refresh', :id => '1'
     should_route 'GET /sla/caches/refresh' => 'sla_caches#refresh'
     should_route 'GET /sla/caches/purge' => 'sla_caches#purge'
+
+    should_route 'GET /issues/1/sla_explain' => 'sla_caches#explain', :id => '1'
   end
 
 end